refactor: add labels as model config
