+++ packages/ai-lab/src/components/AILabImage/AILabImage.tsx
@@ -9,6 +9,7 @@ import {
   predictClassification,
   predictSSD,
 } from '../../lib/helpers';
+import { CLASSES } from '../labels';
 
 const defaultModelConfig: ModelConfig = {
   modelType: 'ssd',
@@ -17,6 +18,7 @@ const defaultModelConfig: ModelConfig = {
   iouThreshold: 0.5,
   nmsActive: true,
   topK: 5,
+  labels: CLASSES,
 };
 
 export const AILabImage = ({
@@ -101,7 +103,7 @@ export const AILabImage = ({
     (async function () {
       if (results) {
         const detections = await getModelDetections(results, modelConfig);
-        const inferences = await getInferenceData(detections);
+        const inferences = await getInferenceData(detections, modelConfig);
         setDetectionResults(detections);
         onInference?.(inferences);
       }
