+++ packages/ai-lab/src/components/AILabVideo/AILabLocalVideo.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useRef, useState } from 'react';
 import * as tf from '@tensorflow/tfjs';
 import { Performance, perfInfo, PerformanceInfo } from '../../performance';
-import { CLASSES } from '../../lib/labels';
+import { CLASSES } from '../labels';
 import { VideoProps } from '../../types';
 
 let activeInfer = false;
